refactor: extract host relay transport to common
Move the CSP relay hook out of like-utils so a later zap button can share the same host seam.
